pyproject.toml
[tool.black]
include = '\.py$'
# Don't attempt to normalize strings
# (Preserves the convention of "single quotes for data, double quotes for humans")
skip-string-normalization = true

[tool.isort]
profile = "black"

[tool.poetry]
name = "ws"
version = "0.1.0"
description = "The MIT Outing Club's trip management system"
license = "GPL-3.0"
readme = "README.md"
homepage = "https://mitoc-trips.mit.edu"
repository = "https://github.com/DavidCain/mitoc-trips/"
authors = ["David Cain <davidjosephcain@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.2"
Django = "^2.2"
PyJWT = ">= 2.0.0"  # 2.0.0 changed the return type of `jwt.encode()`
# 4.4.4 had a KeyError from the `future` module
# 5.0 drops support for the AMQP backend, which we currently use
celery = "> 4.4.4, <5.0"
python-dateutil = "*"  # TODO (Can remove once on Python 3.7)
django-allauth = "^0.41.0"  # See CVE-2019-19844
django-angular = "^2.1"
django-cors-headers = "*"
django-localflavor = "*"
django-phonenumber-field = "<2.4.0"  # Version 2.4.0 affects `cell_phone` fixture
django-pipeline = "*"  # TODO: To eventually be replaced by webpack-loader
django-smtp-ssl = "*"
django-webpack-loader = "*"
gspread = "^3.0.1"
markdown2 = "*"
mitoc-const = "^1.0.0"  # (1.0.0 includes type hints)
mysqlclient = "^1.3.14"  # TODO: Delete once `util/geardb.py` is migrated
oauth2client = "*"
phonenumberslite = "*"
psycopg2-binary = "*"
pwned-passwords-django = "^1.4"
requests = "*"
sentry-sdk = "*"

# Optional dependencies for local development
# (Included here, since extras cannot refer to dev dependencies)
django-debug-toolbar = { version = "*", optional = true }
ipdb = { version = "*", optional = true }
jedi = { version = "<0.18.0", optional = true }  # https://github.com/ipython/ipython/issues/12740

# Optional dependencies for generating & reporting test coverage stats
coverage = { version = "*", optional = true }

# Optional dependencies for static type checking
django-stubs = { version = "*", optional = true }
mypy = { version = "*", optional = true }
types-certifi = { version = "*", optional = true }
types-cryptography = { version = "*", optional = true }
types-python-dateutil = { version = "*", optional = true }
types-pytz = { version = "*", optional = true }
types-requests = { version = "*", optional = true }


[tool.poetry.extras]
code_coverage = ["coverage", "codecov"]
debugging = ["django-debug-toolbar", "ipdb", "jedi"]
typing = [
    "django-stubs",
    "mypy",
    "types-certifi",
    "types-cryptography",
    "types-python-dateutil",
    "types-pytz",
    "types-requests"
]

[tool.poetry.dev-dependencies]
beautifulsoup4 = "*"
black = { version = "*", allow-prereleases = true }
factory_boy = "*"
freezegun = "*"
isort = { version = "^5.8", extras = ["pyproject"] }  # Version 5 introduces black compatibility
lxml = "*"
pylint = "^2.8"
pytz = "*"  # (It's a production dependency by way of Django, but used directly in tests)
